breaks logic into separate files
